requirements.txt
pyyaml
python-box
Flask
langchain 
langchain-text-splitters 
langchain-community
langchain[google-genai]
bs4
python-dotenv
pypdf
langchain-huggingface
sentence-transformers
torch
transformers
-e .